+++ AnimalData/Model/Amphibian.cs
@@ -4,7 +4,7 @@ namespace AnimalData.Model
 {
     internal class Amphibian : TableAnimal
     {
-        public Amphibian(int id, string animalName, byte lifeExpectancy, int weight, string animalType) : base(id, animalName, lifeExpectancy, weight, animalType)
+        public Amphibian(int id, string animalName, byte lifeExpectancy, double weight, string animalType) : base(id, animalName, lifeExpectancy, weight, animalType)
         {
         }
     }
