+++ AnimalData/Model/Bird.cs
@@ -4,7 +4,7 @@ namespace AnimalData.Model
 {
     internal class Bird : TableAnimal
     {
-        public Bird(int id, string animalName, byte lifeExpectancy, int weight, string animalType) : base(id, animalName, lifeExpectancy, weight, animalType)
+        public Bird(int id, string animalName, byte lifeExpectancy, double weight, string animalType) : base(id, animalName, lifeExpectancy, weight, animalType)
         {
         }
     }
