+++ AnimalData/Model/Mammal.cs
@@ -4,7 +4,7 @@ namespace AnimalData.Model
 {
     internal class Mammal : TableAnimal
     {
-        public Mammal(int id, string animalName, byte lifeExpectancy, int weight, string animalType) : base(id, animalName, lifeExpectancy, weight, animalType)
+        public Mammal(int id, string animalName, byte lifeExpectancy, double weight, string animalType) : base(id, animalName, lifeExpectancy, weight, animalType)
         {
         }
     }
